backend/requirements.txt
fastapi
uvicorn[standard]
pillow
rembg
python-multipart
